Creating first Array of Encrypted Ballots
This COmmit brings teh first creation of a struct similar to the one
used in the encryptedBallot on voter. THis is only an intermediary
commiCreating first Array of Encrypted Ballots

This COmmit brings teh first creation of a struct similar to the one
used in the encryptedBallot on voter. THis is only an intermediary
committ